Fix issue with adapters, add meetings metadata to items page
Without this commit, most important adapters inherit from the collections adapter. This commit moves the behavior in the collections adapter (which is important to basically every other adapter) into the application adapter to make the adapter hierarchy more intuitive. Also, this commit adds a section to the item detail page that contains location and time of the presentation.