Make changes on frontend to reflect API collection model refactor
As described in this commit's back-end sibling, this commit aligns the frontend with changes made to the backend which address scalability and flexibility in terms of adding new collection and item types. This commit also removes unusable collection types from the New Collection form, to be added back in once those collection types are implemented.