Fix display bug on collection edit form
Meeting information was not editable and type was not loading. With this commit, both of those issues are fixed. The issues are caused by variable accesses not being updated when we switched from storing collection type in the settings JSON to storing it directly on the model. This updates those accesses.